+++ src/Forms/Components/JsonInput.php
@@ -7,12 +7,12 @@ use CodebarAg\FilamentJsonField\Concerns\HasDarkTheme;
 use CodebarAg\FilamentJsonField\Concerns\HasFoldingCode;
 use CodebarAg\FilamentJsonField\Concerns\HasLineNumbers;
 use CodebarAg\FilamentJsonField\Concerns\HasLineWrapping;
-use Filament\Forms\Components\Concerns;
+use Filament\Forms\Components\Concerns\CanBeReadOnly;
 use Filament\Forms\Components\Field;
 
 class JsonInput extends Field
 {
-    use Concerns\CanBeReadOnly;
+    use CanBeReadOnly;
     use HasAutoCloseBrackets;
     use HasDarkTheme;
     use HasFoldingCode;
@@ -20,6 +20,11 @@ class JsonInput extends Field
     use HasLineWrapping;
 
     protected string $view = 'filament-json-field::forms.components.json-input';
+
+    public static function getAssetPackageName(): ?string
+    {
+        return 'laravel-filament-json-field';
+    }
 
     public function setUp(): void
     {
@@ -30,5 +35,10 @@ class JsonInput extends Field
             ->validationMessages([
                 'array' => __('The :attribute must be valid JSON.'),
             ]);
+    }
+
+    public function getStatePath(bool $isAbsolute = true): ?string
+    {
+        return parent::getStatePath($isAbsolute);
     }
 }