+++ CHANGELOG.md
@@ -1,3 +1,16 @@
 # Changelog
 
 All notable changes to `filament-json-field` will be documented in this file.
+
+## [13.0.0] - 2024-12-XX
+
+### Added
+- Support for Filament 4.x
+
+### Changed
+- Updated dependencies to support Filament 4.x
+- Updated Blade components to use new Filament 4 syntax
+- Updated service provider to use Filament 4 asset registration
+
+### Removed
+- Support for Filament 3.x (use v12.0.0 for Filament 3.x compatibility)